# backend-api/openapi.yaml
openapi: 3.0.3
info:
  title: Eコマース API
  description: AWS オブザーバビリティ学習のための Eコマースアプリケーション API
  version: 0.1.0
  contact:
    name: Your Name
    email: your.email@example.com
servers:
  - url: http://backend-api.localhost/api
    description: ローカル開発サーバー
paths:
  /health:
    get:
      summary: サービスヘルスチェック
      description: APIサービスの稼働状態と基本的な健全性情報を提供します
      operationId: healthCheck
      tags:
        - system
      parameters:
        - name: checks
          in: query
          description: "実行する詳細チェック（カンマ区切り、例: db,iam,s3）"
          required: false
          schema:
            type: string
            example: db,iam,s3
      responses:
        "200":
          description: サービスは正常に稼働しています
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products:
    get:
      summary: 商品一覧の取得
      description: オプションのフィルタリングを使用した、ページネーション対応の商品一覧を返します
      operationId: listProducts
      tags:
        - product
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/CategoryIdParam"
        - $ref: "#/components/parameters/KeywordParam"
      responses:
        "200":
          description: 成功した操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "400":
          description: 無効なパラメータ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}:
    get:
      summary: IDによる商品の取得
      description: 指定されたIDの単一商品を返します
      operationId: getProductById
      tags:
        - product
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: 成功した操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: 商品が見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}/image:
    post:
      summary: 商品画像のアップロード
      description: 指定された商品IDに画像をアップロードします（管理者のみ）
      operationId: uploadProductImage
      tags:
        - product
        - admin
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: アップロードする画像ファイル
              required:
                - image
      responses:
        "200":
          description: 画像が正常にアップロードされました
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                    description: アップロードされた画像のURL
                  thumbnailUrl:
                    type: string
                    description: サムネイル画像のURL
                  mediumUrl:
                    type: string
                    description: 中サイズ画像のURL
                  largeUrl:
                    type: string
                    description: 大サイズ画像のURL
        "400":
          description: 無効なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 権限エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: 商品画像の取得
      description: 指定された商品IDの画像を取得します
      operationId: getProductImage
      tags:
        - product
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
        - name: size
          in: query
          description: 画像サイズ（thumbnail, medium, large）
          schema:
            type: string
            enum: [thumbnail, medium, large]
            default: medium
      responses:
        "200":
          description: 画像が正常に取得されました
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: 画像が見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      summary: カテゴリー一覧の取得
      description: 商品カテゴリーの一覧を返します
      operationId: listCategories
      tags:
        - category
      responses:
        "200":
          description: 成功した操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryList"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories/{id}/products:
    get:
      summary: カテゴリー別商品の取得
      description: 特定のカテゴリーに属する商品のページネーション対応リストを返します
      operationId: listProductsByCategory
      tags:
        - category
      parameters:
        - $ref: "#/components/parameters/CategoryIdPathParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 成功した操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "404":
          description: カテゴリーが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: 予期しないエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    PageParam:
      name: page
      in: query
      description: ページ番号（1ベース）
      schema:
        type: integer
        default: 1
        minimum: 1

    PageSizeParam:
      name: pageSize
      in: query
      description: 1ページあたりのアイテム数
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

    CategoryIdParam:
      name: categoryId
      in: query
      description: カテゴリIDによる商品フィルタリング
      schema:
        type: integer
        format: int

    ProductIdParam:
      name: id
      in: path
      description: 商品ID
      required: true
      schema:
        type: integer
        format: int

    CategoryIdPathParam:
      name: id
      in: path
      description: カテゴリーID
      required: true
      schema:
        type: integer
        format: int

    KeywordParam:
      name: keyword
      in: query
      description: 商品名や説明に含まれる検索キーワード
      required: false
      schema:
        type: string

  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - categoryId
        - inStock
      properties:
        id:
          type: integer
          format: int
          description: 商品の一意識別子
        name:
          type: string
          description: 商品名
        description:
          type: string
          description: 商品説明
        sku:
          type: string
          description: SKUコード
        price:
          type: number
          format: float
          description: 商品価格
        salePrice:
          type: number
          format: float
          description: セール価格
        imageUrl:
          type: string
          description: 商品画像のURL
        inStock:
          type: boolean
          description: 在庫があるかどうか
        stockQuantity:
          type: integer
          description: 在庫数
        categoryId:
          type: integer
          format: int
          description: カテゴリー識別子
        categoryName:
          type: string
          description: カテゴリー名
        isNew:
          type: boolean
          description: 新着商品かどうか
        isFeatured:
          type: boolean
          description: おすすめ商品かどうか
        createdAt:
          type: string
          format: date-time
          description: 作成タイムスタンプ
        updatedAt:
          type: string
          format: date-time
          description: 最終更新タイムスタンプ

    ProductList:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
          description: 商品リスト
        total:
          type: integer
          description: 商品の総数
        page:
          type: integer
          description: 現在のページ番号
        pageSize:
          type: integer
          description: 1ページあたりのアイテム数
        totalPages:
          type: integer
          description: 総ページ数

    Category:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: integer
          format: int
          description: カテゴリーの一意識別子
        name:
          type: string
          description: カテゴリー名
        slug:
          type: string
          description: カテゴリーのURL用スラッグ（一意）
        description:
          type: string
          nullable: true
          description: カテゴリーの説明
        imageUrl:
          type: string
          nullable: true
          description: カテゴリー画像のURL
        parentId:
          type: integer
          format: int
          nullable: true
          description: 親カテゴリーのID（階層構造用）
        productCount:
          type: integer
          description: このカテゴリーの商品数

    CategoryList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
          description: カテゴリーリスト

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
        - uptime
        - resources
        - components
      properties:
        status:
          type: string
          description: サービスの状態
          example: "ok"
        version:
          type: string
          description: APIのバージョン
          example: "0.1.0"
        timestamp:
          type: string
          format: date-time
          description: レスポンスタイムスタンプ
          example: "2024-03-14T12:34:56Z"
        uptime:
          type: integer
          format: int64
          description: サーバーの稼働時間（ミリ秒）
          example: 86400500
        resources:
          $ref: "#/components/schemas/SystemResources"
        components:
          type: object
          description: 各種依存コンポーネントの状態
          additionalProperties:
            type: string
          example:
            api_server: ok
            database: ok
            iam_auth: ok
            s3_connectivity: "error: timeout"

    SystemResources:
      type: object
      description: システムリソースの状態
      additionalProperties:
        type: object
        properties:
          memory:
            $ref: "#/components/schemas/MemoryStats"
          goroutines:
            type: integer
            description: 実行中のゴルーチン数
        required:
          - memory
          - goroutines
      example:
        system:
          memory:
            allocated: 8192000
            total: 16384000
            system: 32768000
          goroutines: 10

    MemoryStats:
      type: object
      properties:
        allocated:
          type: integer
          format: uint64
          description: 割り当てられたメモリ量（バイト）
        total:
          type: integer
          format: uint64
          description: 合計割り当てメモリ量（バイト）
        system:
          type: integer
          format: uint64
          description: システムメモリ使用量（バイト）
      required:
        - allocated
        - total
        - system

    DependentServices:
      type: object
      description: 依存サービスの状態
      additionalProperties:
        $ref: "#/components/schemas/ServiceStatus"
      example:
        api:
          name: "aws-observability-ecommerce"
          status: "up"

    ServiceStatus:
      type: object
      properties:
        name:
          type: string
          description: サービス名
        status:
          type: string
          description: サービスの状態（up/downなど）
      required:
        - name
        - status

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: "invalid_parameter"
        message:
          type: string
          description: エラーメッセージ
          example: "Invalid parameter value"
        details:
          type: object
          additionalProperties: true
          description: 追加のエラー詳細
          example:
            field: "page"
            value: "-1"
            reason: "must be greater than or equal to 1"
        timestamp:
          type: string
          format: date-time
          description: エラー発生タイムスタンプ
          example: "2024-03-14T12:34:56Z"
        traceId:
          type: string
          description: デバッグ用のリクエストトレースID
          example: "abc123def456"

tags:
  - name: system
    description: システムの状態や健全性に関するAPI。サーバーのヘルスチェック、リソース情報、メトリクスなどを含む。
  - name: product
    description: 商品カタログ関連のAPI。商品一覧の取得、商品詳細、カテゴリー情報など、eコマースの主要商品データにアクセスするためのエンドポイント。